feat: working commands except interrupts

+++ Core/Inc/queue.h
@@ -15,9 +15,9 @@
 
 struct fifo_queue {
 	size_t data_p;
-	size_t counter;
+	int counter;
 	uint8_t data[MAX_QUEUE_SIZE];
-	size_t line_feeds;
+	int line_feeds;
 };
 
 void queue_write(struct fifo_queue*, uint8_t*, size_t);
